pyproject.toml
[tool.black]
target-version = ['py39']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | \venv
    | \.aws-sam
    | _build
    | buck-out
    | build
    | dist
    | node_modules
  )/
)
'''

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.semantic_release]
version_variable = [
    "django_microsoft_sso/__init__.py:__version__",
    "pyproject.toml:version"
]
branch = "main"
upload_to_pypi = true
upload_to_release = true
build_command = "python -m pip install -U twine poetry && poetry build"

[tool.poetry]
name = "django-microsoft-sso"
version = "0.1.0"
description = "Easily add Microsoft Authentication to your Django Projects"
authors = ["Chris Maillefaud <chrismaille@users.noreply.github.com>"]
readme = "README.md"
repository = "https://github.com/megalus/django-microsoft-sso"
keywords = ["microsoft", "django", "sso"]
license = "MIT"
classifiers = [
    "Framework :: Django",
    "Framework :: Django :: 4.0",
    "Framework :: Django :: 4.1",
    "Framework :: Django :: 4.2",
    "Intended Audience :: Developers",
    "Development Status :: 4 - Beta",
    "Environment :: Plugins"
]

[tool.poetry.dependencies]
python = ">=3.10, <4.0"
django = ">=4.0"
loguru = "*"
identity = "*"
httpx = "*"

[tool.poetry.dev-dependencies]
auto-changelog = "*"
black = {version = "*", allow-prereleases = true}
pre-commit = "*"
pytest-coverage = "*"
pytest-django = "*"
pytest-lazy-fixture = "*"
pytest-mock = "*"
twine = "*"
python-dotenv = "*"
mkdocs-material = "*"
django-grappelli = "*"
django-jazzmin = "*"
django-admin-interface = "*"
django-jet-reboot = "*"
click = ">8"
bandit = "*"
flake8 = "*"
stela = "*"
django-google-sso = "*"
# django-google-sso = {path = "../django-google-sso", develop = true}
httpx = "*"

[tool.stela]
environment_variable_name = "STELA_ENV"
evaluate_data = true
show_logs = false
dotenv_overwrites_memory = true
env_file = ".env"
config_file_path = "./example_app"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
